Finish mining function

+++ db/migrate/20180221225918_create_coins.rb
@@ -3,9 +3,9 @@ class CreateCoins < ActiveRecord::Migration[5.1]
     create_table :coins do |t|
       t.string :name, default: 'Bitcoin'
       t.string :symbol, default: "BTC"
-      t.float :difficulty, default: 0.0
-      t.float :amount, default: 0.0
-      t.float :price, default: 0.0
+      t.decimal :difficulty, default: 0.0
+      t.decimal :amount, default: 0.0
+      t.decimal :price, default: 0.0
 
       t.timestamps
     end
