+++ db/migrate/20180221234354_create_miners.rb
@@ -1,8 +1,8 @@
 class CreateMiners < ActiveRecord::Migration[5.1]
   def change
     create_table :miners do |t|
-      t.float :consumption, default: 0.0
-      t.float :hash_rate, default: 0.0
+      t.decimal :consumption, default: 0.0
+      t.decimal :hash_rate, default: 0.0
 
       t.timestamps
     end
