+++ db/migrate/20180224231646_create_energy_packs.rb
@@ -2,7 +2,7 @@ class CreateEnergyPacks < ActiveRecord::Migration[5.1]
   def change
     create_table :energy_packs do |t|
       t.integer :size, default: 15600
-      t.float :price, default: 100.00
+      t.decimal :price, default: 100.00
       t.timestamps
     end
   end
