+++ db/migrate/20180225010131_add_balance_to_user.rb
@@ -1,5 +1,5 @@
 class AddBalanceToUser < ActiveRecord::Migration[5.1]
   def change
-    add_column :users, :balance, :float, default: 2000.0
+    add_column :users, :balance, :decimal, default: 2000.0
   end
 end
